Change post-increment to pre
For some reason in bash arithmetic expansion when the returned value is
0 the exit code will be 1 - otherwise the exit code is 0. With post
increment the current value will be returned then the value in memory
will be incremented. This means that post-incrementing starting at 0
will have a 1 exit code on the first increment. set -e treats this as an
error.

Pre incrementing returns the value after incrementing so there will
never be a 0 return value when incrementing from 0.

Signed-off-by: Colin Simmons <colin.simmons@engineerbetter.com>